+++ backend/src/main/java/com/lambdaschool/coffeebean/config/ResourceServerConfig.java
@@ -26,10 +26,13 @@ public class ResourceServerConfig extends ResourceServerConfigurerAdapter
         http.
                 anonymous().disable()
                 .authorizeRequests()
-                .antMatchers("/orders/**").access("hasAnyRole('ROLE_ADMIN', 'ROLE_USER')")
+                .antMatchers("/orders/**").access("hasAnyRole('ROLE_ADMIN')")
                 .antMatchers("/products/**").access("hasAnyRole('ROLE_ADMIN')")
-                .antMatchers("/suppliers/**").access("hasAnyRole('ROLE_USER')")
-//                .antMatchers("/users/**").access("hasAnyRole('ROLE_ADMIN', 'ROLE_USER')")
+                .antMatchers("/suppliers/**").access("hasAnyRole('ROLE_ADMIN')")
+                .antMatchers("/users/**").access("hasAnyRole('ROLE_ADMIN')")
+
+                .antMatchers("/cart/**").access("hasAnyRole('ROLE_ADMIN', 'ROLE_USER')")
+
                 .and().exceptionHandling().accessDeniedHandler(new OAuth2AccessDeniedHandler());
     }
 }
