+++ backend/src/main/java/com/lambdaschool/coffeebean/config/SecurityConfig.java
@@ -83,7 +83,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter
                 .authorizeRequests()
                 .antMatchers("/swagger-ui.html").permitAll()
                 .antMatchers("/api-docs/**").permitAll()
-                .antMatchers("/users/**").permitAll()
+//                .antMatchers("/users/**").permitAll()
 //                .antMatchers("/console/**").permitAll()
                 .antMatchers(HttpMethod.POST, SIGN_UP_URL).permitAll()
                 .antMatchers(HttpMethod.GET, SHOP_URL).permitAll();
