+++ backend/src/main/java/com/lambdaschool/coffeebean/controller/Ordercontroller.java
@@ -2,7 +2,6 @@ package com.lambdaschool.coffeebean.controller;
 
 import com.lambdaschool.coffeebean.model.Order;
 import com.lambdaschool.coffeebean.repository.Orderrepository;
-import com.lambdaschool.coffeebean.repository.Userrepository;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
@@ -22,9 +21,6 @@ import java.util.Optional;
 public class Ordercontroller {
     @Autowired
     Orderrepository orderrepos;
-
-    @Autowired
-    Userrepository userrepos;
 
     @ApiOperation(value = "find all orders - DKM", response = Order.class)
     @ApiResponses(value = { @ApiResponse(code = 200, message = "Successfully received customer - DKM"),
@@ -79,12 +75,5 @@ public class Ordercontroller {
     // }
     // }
 
-    // ============ Add To Order ================
 
-    @PostMapping("/buy/{userid}")
-    public Order buyItemsInCart(@RequestBody Order newOrder, @PathVariable long userid) {
-        userrepos.deleteAllItemsFromCart(userid);
-
-        return orderrepos.save(newOrder);
-    }
 }
